Make cef-ui-helper a library and a binary
